Make also fail in the CI

+++ tests/regression/73-strings/09-malloc.c
@@ -1,5 +1,4 @@
 // PARAM: --disable ana.base.limit-string-addresses --enable ana.int.interval --enable ana.base.arrays.nullbytes
-
 #include <goblint.h>
 #include <string.h>
 #include <stdlib.h>
@@ -11,6 +10,7 @@ int main () {
     char s2[50];
     s2[0] = 'a';
 
-    int len1 = strlen(s1); //WARN
-    int len2 = strlen(s2); //WARN
+    // Use size_t to avoid integer warnings hiding the lack of string warnings
+    size_t len1 = strlen(s1); //WARN
+    size_t len2 = strlen(s2); //WARN
 }
